Move GenreManager to it's own file
